feat: adding useReducer done. Ref sn
