+++ apps/05-hook-app/08-useReducer/TodoApp.tsx
@@ -1,13 +1,9 @@
 import "./styles.css"
-import { FormEvent, useEffect, useReducer } from "react"
+import { useEffect, useReducer } from "react"
 import { Todo } from "./types"
-import todoReducer, { TodoReducerAction } from "./todoReducer"
-import useForm from "../02-useEffect/hooks/useForm"
+import todoReducer from "./todoReducer"
 import TodoList from "./components/TodoList"
-
-interface FormValues {
-  description: string
-}
+import TodoAdd from "./components/TodoAdd"
 
 const initialState: Todo[] = []
 
@@ -19,9 +15,6 @@ const init = () => {
 
 const TodoApp = () => {
   const [todos, dispatch] = useReducer(todoReducer, initialState, init)
-  const [formValues, handleInputChanges, resetFormValue] = useForm<FormValues>({
-    description: "",
-  })
 
   useEffect(() => {
     localStorage.setItem("todos", JSON.stringify(todos))
@@ -35,24 +28,11 @@ const TodoApp = () => {
     dispatch({ type: "toggle", payload: id })
   }
 
-  const handleAddTodo = (e: FormEvent) => {
-    e.preventDefault()
-
-    if (formValues.description.trim().length === 0) {
-      return
-    }
-
-    const newTodo: Todo = {
-      id: new Date().getTime().toString(),
-      description: formValues.description,
-      done: false,
-    }
-    const action: TodoReducerAction = {
+  const handleAddTodo = (newTodo: Todo): void => {
+    dispatch({
       type: "add",
       payload: newTodo,
-    }
-    dispatch(action)
-    resetFormValue()
+    })
   }
 
   return (
@@ -71,27 +51,7 @@ const TodoApp = () => {
           />
         </div>
         <div className={"col-5"}>
-          <h4>Add Todo</h4>
-          <hr />
-
-          <form action="" onSubmit={handleAddTodo}>
-            <input
-              type="text"
-              name={"description"}
-              className={"form-control"}
-              placeholder={"learn..."}
-              autoComplete={"off"}
-              value={formValues.description}
-              onChange={handleInputChanges}
-            />
-
-            <button
-              type={"submit"}
-              className={"btn btn-outline-primary mt-1 btn-block"}
-            >
-              Add
-            </button>
-          </form>
+          <TodoAdd handleAddTodo={handleAddTodo} />
         </div>
       </div>
     </>
